PhotoProcessor/requirements.txt
pika==1.1.0
Pillow==8.1.0
pymongo==3.11.3
